requirements.txt
pytest
pytest-cov
numpy
pyserial
